Refactors CheckoutSDK initialization and result handling
Updates the CheckoutSDK to use a `start` method for initialization, separating it from the constructor. This change improves flexibility and allows for deferred configuration.

Implements success and error handling within the CheckoutViewController, which now dismisses the view controller and triggers the corresponding `onSuccess` or `onError` methods on the CheckoutSDKResults delegate.

Moves CheckoutViewController to the Private directory.